+++ server/py/services/api/crud/runtimes/nuclio/function.py
@@ -23,6 +23,7 @@ import nuclio.utils
 import requests
 
 import mlrun
+import mlrun.auth.nuclio
 import mlrun.auth.utils
 import mlrun.common.constants
 import mlrun.common.constants as mlrun_constants
@@ -91,7 +92,7 @@ def deploy_nuclio_function(
             create_new=mlrun.mlconf.httpdb.projects.leader == "mlrun",
             watch=False,
             return_address_mode=nuclio.deploy.ReturnAddressModes.all,
-            auth_info=auth_info.to_nuclio_auth_info() if auth_info else None,
+            auth_info=mlrun.auth.nuclio.NuclioAuthInfo.from_auth_info(auth_info),
         )
     except nuclio.utils.DeployError as exc:
         if exc.err:
@@ -161,7 +162,7 @@ def get_nuclio_deploy_status(
             verbose,
             resolve_address,
             return_function_status=True,
-            auth_info=auth_info.to_nuclio_auth_info() if auth_info else None,
+            auth_info=mlrun.auth.nuclio.NuclioAuthInfo.from_auth_info(auth_info),
         )
     except requests.exceptions.ConnectionError as exc:
         mlrun.errors.raise_for_status(
